Move window and shader to a graphics directory
